requirements.txt
yfinance>=0.2.44
pandas>=2.2.0
pyarrow>=14.0.0
numpy>=1.26.0
plotly>=5.24.0
scipy>=1.13.0
akshare>=1.12.0
schedule>=1.2.0

# API service
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
websockets>=12.0

# ML models
scikit-learn>=1.4.0
scikit-optimize>=0.10.2
xgboost>=2.0.0
lightgbm>=4.3.0
optuna>=3.0.0

# Time-series forecasting
prophet>=1.1.5
statsmodels>=0.14.0

# Optional deep-learning models
# torch>=2.1.0
# torchvision>=0.16.0
# tensorflow>=2.15.0
# keras>=3.0.0
# pytorch-forecasting>=1.0.0
# darts>=0.27.0

# Optional data providers
# tushare>=1.2.0
# ccxt>=4.0.0

# Security and auth
bcrypt>=4.0.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0

# Monitoring
psutil>=5.9.0
requests>=2.31.0
aiohttp>=3.9.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-asyncio>=0.21.0
pytest-benchmark>=4.0.0
pytest-playwright>=0.4.0
httpx>=0.24.0

# LLM / search
openai>=1.0.0
google-generativeai>=0.3.0
tavily-python>=0.3.0
